+++ CHANGELOG.md
@@ -4,8 +4,10 @@ tomland uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
 ## 1.1.0.1 — Jul 10, 2019
+
 * [#206](https://github.com/kowainik/tomland/issues/206):
-  Bug fix in parser and additional tests
+  Fix in parser of inline tables inside tables, add tests for official TOML examples
+  (by [@jiegillet](https://github.com/jiegillet))
 
 ## 1.1.0.0 — Jul 8, 2019